refactor: Filter audio instruction links by file extension
This commit refactors the code in OrderOrderItemCommonService.java to filter the audio instruction links by file extension. Specifically, it modifies the line that retrieves the audio instruction links to only include links that end with ".mp3". This change ensures that only MP3 audio files are included in the list of audio instruction links.