Stg/ashu/audio_instruction_item_pdf_issue (#61)
* feat: Add audio instructions to item details page

This commit adds the functionality to display audio instructions on the item details page. If there are any audio instructions available, they will be shown in a separate section with an audio player for each instruction. If the browser does not support the audio element, a fallback message with a link to listen to the audio externally is provided.

* refactor: Filter audio instruction links by file extension

This commit refactors the code in OrderOrderItemCommonService.java to filter the audio instruction links by file extension. Specifically, it modifies the line that retrieves the audio instruction links to only include links that end with ".mp3". This change ensures that only MP3 audio files are included in the list of audio instruction links.

* feat: Improve audio instructions display on item details page

---------

Co-authored-by: raghav2499 <mittalr0204@gmail.com>
Co-authored-by: Raghav Mittal <justcrack1206@gmail.com>